fix warnings

+++ tests/tests/starknet/keccak.rs
@@ -287,7 +287,7 @@ impl StarknetSyscallHandler for SyscallHandler {
         todo!()
     }
 
-    fn set_max_fee(&mut self, _max_fee: u128) -> SyscallResult<()> {
+    fn set_max_fee(&mut self, _max_fee: u64) -> SyscallResult<()> {
         todo!()
     }
