+++ tests/tests/starknet/secp256.rs
@@ -272,7 +272,7 @@ impl StarknetSyscallHandler for &mut SyscallHandler {
         todo!()
     }
 
-    fn set_max_fee(&mut self, _max_fee: u128) -> SyscallResult<()> {
+    fn set_max_fee(&mut self, _max_fee: u64) -> SyscallResult<()> {
         todo!()
     }
